in mts, sort outgoing moves by creation date
This change should avoid circular situations where the plan finishes and
some outgoing moves are untouched, changing the order the next time that
the dates are recomputed. In that case the operation was not atomic,
something we would like to avoid.

Now the moves are sorted by creation date (which does not change) and at
the end the expected date is written. That way we should get always the
same results the next time.