Started stock_picking_pallet module
Allows to generate a pallet for a selection of pickings

Implements the generation of the package for the selected pickings

Add a sequence for the pallets

Add security for model stock.picking.pallet

Add the base files for the pallet report

Do no longer use a sequence for the name

Use the related pack name instead

Do not copy some fields on duplicate

Rename dest_pack_id to dest_package_id

Report: show infos and lines

Not really nicely formatted though. Some info are missing

Add weights and use a field for quants

Split in 2 steps: put in pack, set to done

The report can only be printed once it is done, because it needs the packs tree
to be created.

The life_date field depends on product_expiry

Improve layout

Add date_done when a pallet is done

Change conditions to hide fields

Rename dest_package_id to package_id

Rename the model the stock.picking.package.preparation

stock.picking.pallet was too specific, this module can be used
to generate any kind of package.

Rename files from pallet to package_preparation

Explicit the relations otherwise the name is too long

Remove the report: too specific

Add tests on preparation of the package

Add help messages along the way

Add a test on the weight

Improve the readme with the OCA's layout

Add Odoo Community Association (OCA) as author

Add translation pot

Add French translation

Allow the action_* methods to run on several records

Do not repeat the dict for the fields' states