Unreserve quants when blocking lots
Users block lots when there is a problem with it (eg. quality issue), and it's important that no delivery is made when such a problem is found. So blocking is more important than reservations.
So, ask for user confirmation when blocking, and then break the reservations.
Includes translation into French