[ADD] shipping_expedition: New module to manage shipments made through external transport companies
This module allows you to create a new model called: shipping.expedition that is integrated with the Inventory application in order to manage external shipments that are made by external transport companies.
The creation of shipments is allowed from the model: stock.picking according to the carrier defined in it.

It is a base functionality that through other addons will allow integration with different transport companies: Nacex (http://www.nacex.es/), CBL (https://www.cbl-logistica.com/) y TXT (https://txt.es/)
A common model is generated and its data will be used to connect to the corresponding webservice and / or generate the .txt file that the programs need.
Additionally (with the additional addons) the status of the expeditions will be updated.