Add stock_putaway_hook
Add hooks allowing modules to add more putaway strategies.
It adds hooks to the core putaway method
``StockLocation._get_putaway_strategy()`` allowing to plug other
strategies and makes the selector fields in the tree views dynamic
(required/readonly).

The fields_view_get override is to handle a tricky issue:

By default, the product_id field as such attrs:

"{'readonly': [('category_id', '!=', False)], 'required': [('category_id', '=', False)]}"

And the category_id field:

"{'readonly': [('product_id', '!=', False)], 'required': [('product_id', '=', False)]}"

If we add a new field "route_id", we want to extend the domains to
include it, such as:

"{'readonly': ['|', ('category_id', '!=', False), ('route_id '!=', False)],
  'required': [('category_id', '=', False), ('route_id', '=', False)]}"

And every new field will need to be added in the product_id and
category_id field, but also in the other fields (exemple, if we add a
"foo_id" field, it needs to be added in the attrs of "route_id"), which
isn't possible with the extension mechanisms of the XML views.

The solution provided by this module is that new fields must have the
{'exclusive_selection': True} in the options attribute, and an override
of fields_view_get automatically generates the attrs for all these
"exclusive" fields.

The code here will be tested by `stock_putaway_by_route` in a follow-up
commit.