Add stock_putaway_by_route
Based on stock_putaway_hook.
Add the possibility to apply a putaway rule based on a route.

For instance, a move generated for a replenishment (e.g. using the module
``Stock Orderpoint Route`` or ``Stock Buffer Route``) will be related to the
rule (so the route) that generated it. When the putaway rules are applied, if a
rule is defined on the replenishment route, the destination move will be changed
to the one of the rule.