[13.0][IMP] stock_picking_group_by_partner_by_carrier
The delivery slip can reference more than one sale order with different
customer references. If there is more than one of those references they
will be displayed in the lines.
But only one in the header, better not to display the customer reference
in the header with multiple sale order.
Also fix the customer reference to be specific to the delivery lines.