Fix cancelation of SO when multiple SO are grouped in a same delivery.
Propagate the original_group_id on pull moves to allow to prevent
merging pulled moves. This allows to cancel a SO without canceling
pulled moves from other SO, as we ensure they are not merged.

squash! Fix cancelation of SO when multiple SO are grouped in a same delivery.

Add tests