stock_quant_package_dimension_total_weight_from_packaging
First, we add a new 'estimated_pack_weight' on the package in
'stock_quant_package_dimension' module which is similar to the
'weight' field implemented in the standard 'delivery' module: it is
done on purpose to not depend on 'delivery'. This weight is thus
computed from product weight.

Second, we re-implement this field in
'stock_quant_package_dimension_total_weight_from_packaging' module
(auto_install) which make use of the weight computation mechanism from
'product_total_weight_from_packaging'. The weight here is computed from
the packaging weight.