[FIX] stock_move_manual_lot: operations on multiple move lines at once
When the manual lot is assigned to multiple lines on the same picking at once,
it can be the case that some lots are reassigned between these lines. In this
case we have to prevent those lines from being deleted when they are being
unreserved, otherwise when they receive their manual lot in their turn, they
no longer exist ("Record does not exist or has been deleted").

In this change, we determine beforehand to which move lines this is likely
to happen and prevent their deletion (once in the move line's unlink method,
and once in the stock move's _do_unreserve method). When the move is being
rereserved later in the same call, these zero qty move lines are reused.

Co-authored-by: Stefan Rijnhart <stefan@opener.am>