[FIX] stock_picking_invoice_link: Consider DS pickings
Steps to reproduce the problem:

- Create a sales order with Dropshipping route.
- Confirm the sale order and validate the DS picking.
- Invoice the purchase order.
- Invoice the sales order.

Current behavior:

The customer invoice is not linked to the DS picking

Expected behavior:

It's linked.

That's because current algorithm looks for existing invoices linked to
the picking for discarding double links, but it's not having into
account that such invoices can be vendor ones.

We simply filter for customer invoices in the algorithm to fix it.

TT37997