[FIX] purchase_stock_picking_invoice_link: Incorrect relation when backorders
Before this patch, when we receive products in a backorder and the other
quantities received where invoiced yet, the picking of the backorder
will be related to the invoice created for the other quantities.

With this patch, this situation will just happen when the invoicing
policy of the product is based on quantities purchased.

Steps to reproduce the problem:

1. Create a purchase order buying 10 units of some stockable product
2. Receive 8 units of the product
3. Create back order
4. Invoice the received quantity (8 units)
5. Receive 2 units of the product in the created back order
6. You will see that the back order picking is related to the invoice with 8 units