[FIX] product_cost_price_avco_sync: Avoid infinite loop writing current SVL
The preprocess hook serves for evaluating for the current SVL being
synced if there's something that conditionate its amount/quantity, and
avoiding the regular process, or if it modifies another SVLs.

The problem is that if you need to write the same SVL that is being
processed, it provokes an infinite loop.

With this fix, we return the expected values to be written in the current
SVL, and store it later at the end of the regular process, but being
cached in our SVL dict values for also being taking into account on the
rest of the process.

We also cache the SVL before doing any process for not having
cache-misses.

TT44658