[FIX] stock_picking_inter_warehouse: make compatible with multi-company
Previously the route was only created for the default company, and
various checks and tests were missing in order to make the module
work in a multi-company environment.

Now an inter-warehouse route is created for each company, including on
installation via an init_hook. A small migration script is provided
as well.