[ADD] stock_customer_deposits
This module extends the functionality of the consignment stock to support customer deposits and allows the management of customer deposits in warehouses.
It is possible to activate the use of customer deposits per warehouse. This will allow the use of an operation type and a route to create customer deposits in the warehouse.

MT-3404 @moduon

Co-authored-by: Rafael Blasco <8736623+rafaelbn@users.noreply.github.com>
Co-authored-by: Gelo Joga <147538094+Gelojr@users.noreply.github.com>
Co-authored-by: Jairo Llopis <973709+yajo@users.noreply.github.com>
Co-authored-by: Eduardo de Miguel <1162050+Shide@users.noreply.github.com>