[IMP] stock_picking_group_by_partner_by_carrier: Avoid procument group swap
When the grouping functionality is enabled. A new procurement group is created the first time a pick out is created as merged. Every time moves from a new group are added into the picking, the merged procurement is updated to link the new group. In this way, the group is never updated after its original creation