[FIX] stock_picking_line_sequence This commit is to follow up the updating behavior of the odoo standard done by this PR(https://github.com/odoo/odoo/pull/144276)
Prior to this update, the integration of this module with the mrp module was problematic.
The core of the issue lies in the renaming of get_aggregated_properties to _get_aggregated_properties and
the modification of the line_key value through inheritance in the mrp module. As a result, the line_key value
did not align with the existing approach, leading to a failure in assigning the sequence2.
This commit aims to rectify this integration issue, ensuring that the module functions correctly when used
alongside the mrp module.