[FIX] bad interaction with mrp_subcontracting
Inject the context variable in stock.picking::_action_done.

Since mrp_subcontracting does two rounds of actions in that method when
a backorder picking is created, we need to add the list of moves for which
propagation was done before the first move is updated, so that the
second round of propagation does not end up updating the same moves
twice.