[IMP] stock_picking_group_by_partner_by_carrier: add failing test for sale link
When a merged procurement group is created, it should not be linked to the
original sales order. Linking it would cause any moves manually added to the
delivery picking to be associated with that sales order.

This commit reproduce the bug and show the manual add of the move impact the sale order