[FIX] mto_as_mts_orderpoint_product_variant: Refactor for archival
There were multiple issues:

- Inverse function on storable field is not reliable and we should
prefer overriding write to add any processing when the field value
is changed. Moreover, the field is_mto on product.product did not
have its inverse function set on the field, so the function was not
called.

- When mto_as_mts is changed on the warehouse, we do not want to
archive MTO rule for this warehouse (because we could still want
to do MTO but not "as MTS").