[REF] stock_restrict_lot: refactor test for compatibility
Small change to a test. In certain cases, other modules may introduce
data, including demo data, that creates a failure in the test.

Specifically, we encountered an error when installing `website_sale`
before this module, since in the demo data a Sales Order is created
and confirmed using product `product.product_product_16`

This means that the `outgoing_qty` is already 1 when the test starts,
instead of the expected 0, and the next checks fail.