[REF] stock_split_picking: Split off moves in a new picking
When a transfer is being partially processed by a user and you want to validate
only that, you should create a new split order containing the done lines and
keep remaining lines in the original transfer.

In business terms, a split order represent an extraction of an original order
that is being partially processed. A back order is the remaining of an order
that was not entirely available.

Implements #1882