[FIX] stock_picking_line_sequence: sequence from sale order
The sequence was wrong on deliveries created on sale order confirmation: it was a copy of the sale order line sequence (for example all 10s if the order lines were never reordered). The reason is that, via stock rules, stock moves aren't created directly with a picking, which would trigger _reset_sequence() in the create() override. Instead, they are initially created without picking and then linked to one. To fix this, we call _reset_sequence() in a write() override as well.