[FIX] stock_move_line_qty_picked: Align move quantity with reserved
In the previous version of this module the computation of
stock.move.line.quantity was modified to reflect the picked quantity.

However, stock.move.line.quantity must reflect the reserved quantity
on the move lines since there is an inverse function that is going
to increase or decrease the quantity on stock.move.line.quantity
accordingly.

The stock.move.quantity must only be updated to reflect the sum
of stock.move.line.qty_picked when stock.move._action_done is called,
to avoid the creation of a backorder.