[ADD] stock_picking_batch_planner
This module extends the functionality of batch operations in Odoo to support the direct
creation of waves from batches and to allow you to plan Waves and Batches with a single
click based on the grouping rules defined in the operation type. It also provides smart
buttons to easily access the Waves and Batches linked to the original Batch, improving
traceability and efficiency in warehouse operations.

MT-11215 @moduon

Co-authored-by: Eduardo de Miguel <1162050+Shide@users.noreply.github.com>