[FIX] stock_no_negative false positive
Because of concurrency, it is possible that a negative quant is created
because the main quant is locked and Odoo will avoid waiting in that
case, and create additional quants for the product in the location.

This fix updates the way the constraint is computed by gathering the
quants and checking the overall quantity.