[FIX] stock_picking_operation_destination_suggestion: labels disappearing in picking type for view.
It seems odoo UI is only showing a label for a field in its view if the field is inside a LEAF <group> node.
The previous xpath inserted the new "Destination Suggestion" group inside an already existing one making the fields of the later not render their labels.

This commit puts the new "Destination suggestion" group AFTER the xpath-selected group (not inside) and adds a name attribute to ensure easier inheritence.